+++ R/TrackTrap-package.R
@@ -1,5 +1,4 @@
 #' @keywords internal
 #' @import dplyr
 #' @import rlang
-#' @import degday
 "_PACKAGE"